+++ lib/Db/TimelineQueryCTE.php
@@ -21,7 +21,9 @@ trait TimelineQueryCTE
         $CLS_TOP_FOLDER = 'f.fileid IN (:topFolderIds)';
 
         // Select 1 if there is a .nomedia file in the folder
-        $SEL_NOMEDIA = "SELECT 1 FROM *PREFIX*filecache f2 WHERE f2.parent = f.fileid AND f2.name = '.nomedia'";
+        $SEL_NOMEDIA = "SELECT 1 FROM *PREFIX*filecache f2
+            WHERE (f2.parent = f.fileid)
+            AND (f2.name = '.nomedia' OR f2.name = '.nomemories')";
 
         // Check no nomedia file exists in the folder
         $CLS_NOMEDIA = "NOT EXISTS ({$SEL_NOMEDIA})";